requirements.txt
numpy==1.18.1
numpydoc==0.9.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pickleshare==0.7.5
Pillow==8.1.0
